requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
backoff==2.2.1
banks==2.4.2
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
filelock==3.29.0
filetype==1.2.0
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
googleapis-common-protos==1.74.0
greenlet==3.5.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.12.2
hyperframe==6.1.0
idna==3.13
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama-index==0.14.21
llama-index-core==0.14.21
llama-index-embeddings-openai==0.6.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.7.7
llama-index-vector-stores-pinecone==0.8.0
llama-index-workflows==2.20.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
mem0ai==2.0.1
mmh3==4.1.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
openai==2.33.0
packaging==24.2
pillow==12.2.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
platformdirs==4.9.6
portalocker==3.2.0
posthog==7.13.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
pypdf==6.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
qdrant-client==1.17.1
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
safetensors==0.7.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.7.0
typer==0.25.0
types-requests==2.33.0.20260408
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
wrapt==2.1.2
yarl==1.23.0
anthropic
google-adk
fastmcp
a2a-sdk==0.3.24
authlib==1.2.1


